# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Облік Майна API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/POST/upload-scan": {
      "post": {
        "summary": "Upload scan image",
        "description": "Upload scan image to server",
        "security": [
          {
            "BearerAuth": [
              "bearer"
            ]
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "./schemas/api-post-uploadScan/requestBody.yaml#/RequestBody"
              },
              "examples": {
                "default": {
                  "$ref": "./schemas/api-post-uploadScan/requestBody-example.yaml"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "filename": {
                      "type": "string",
                      "example": "scan_56.jpeg"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 401
                    }
                  }
                }
              }
            }
          },
          "405": {
            "description": "Method not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "501": {
            "description": "Internal Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Wrong query to database"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}